requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Werkzeug==2.2.2
web3==5.30.0
python-dotenv==1.0.0
